hw8/code/requirements.txt
biopython==1.79
matplotlib==3.1.3
seaborn==0.10.0
numpy==1.20.0
